#60 refactor: 이미지 에셋 경로 관련 상수값 as const
