--- src/constants/imageAddress.ts
@@ -1,11 +1,0 @@
-const IMAGE_ADDRESS = {
-  emailCheckBox: '/assets/yellow/marked.svg',
-  landLoading: '/assets/land/loading.svg',
-  landForm: '/assets/green/search.svg',
-  landTitle: '/assets/green/squiggly.svg',
-  landFriend: '/assets/land/friend.svg',
-  gameClear: '/assets/card/clear.svg',
-  cardFaceDown: '/assets/card/faceDown.svg',
-}
-
-export default IMAGE_ADDRESS
