+++ src/constants/index.ts
@@ -1,5 +1,4 @@
 import { BACKGROUND_TYPE_INFO } from './common/background'
-import IMAGE_ADDRESS from './imageAddress'
 import LOCAL_STORAGE_KEY from './localStorageKey'
 import { LOGO_TYPE_INFO } from './common/logo'
 import { TRIANGLE_TYPE_INFO } from './common/triangle'
@@ -21,7 +20,6 @@ import { LS_EMAIL, LOGIN_INITIAL_VALUES, LOGIN_VALIDATION } from './form/login'
 import { VISIT_VALIDATION, VISIT_INITIAL_VALUE } from './form/visit'
 export {
   BACKGROUND_TYPE_INFO,
-  IMAGE_ADDRESS,
   LOCAL_STORAGE_KEY,
   LOGO_TYPE_INFO,
   TRIANGLE_TYPE_INFO,
