+++ src/types/index.ts
@@ -6,6 +6,7 @@ import {
   triangleTypeInfoProps,
   TriangleProps,
 } from './triangle'
+import { IMAGE_ADDRESS, IMAGE_ADDRESS_KEY } from './styles/imageAddress'
 import { BottomMenuInfoProps } from './bottomMenu'
 import {
   ModalStateTypeInfoProps,
@@ -33,7 +34,7 @@ import {
   playStateActionProps,
 } from './game'
 
-export { playStateAction, GameStatus, GameCardStatus }
+export { playStateAction, GameStatus, GameCardStatus, IMAGE_ADDRESS }
 export type {
   LogoProps,
   logoTypeInfoProps,
@@ -66,4 +67,5 @@ export type {
   GameCardStatusKey,
   playStateActionKey,
   GameStatusKey,
+  IMAGE_ADDRESS_KEY,
 }
